рабивает js код на модули
