+++ index.html
@@ -7,6 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
+  <script src="/js/main.js" type="module"></script>
 </head>
 
 <body>
@@ -233,6 +234,5 @@
   </template>
 
   <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
 </body>
 </html>
