+++ js/data.js
@@ -1,0 +1,76 @@
+import {pickItemFromList, pickIntegerInRange} from './util.js';
+
+const descriptions = [
+  'описание 1',
+  'описание 2',
+  'описание 3'
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const names = [
+  'Артём',
+  'Пётр',
+  'Максим',
+  'Мария',
+  'Дарья',
+  'Екатерина'
+];
+
+/**
+ * создаст объект комментария
+ * @param {number} id
+ * @return {CommentState}
+ */
+const createCommentState = (id) => {
+  const avatar = `img/avatar-${pickIntegerInRange(1, 6)}.svg`;
+  const message = pickItemFromList(messages);
+  const name = pickItemFromList(names);
+
+  return {id, avatar, message, name};
+};
+
+/**
+ * создаст список комментариев
+ * @param {number} length
+ * @return {CommentState[]}
+ */
+const createCommentStateList = (length) => {
+  const list = new Array(length).fill(1);
+
+  return list.map((start, index) => createCommentState(start + index));
+};
+
+/**
+ * создаст объект фотографии
+ * @param {number} id
+ * @return {ImageState}
+ */
+const createImageState = (id) => {
+  const url = `photos/${id}.jpg`;
+  const description = pickItemFromList(descriptions);
+  const likes = pickIntegerInRange(15, 200);
+  const comments = createCommentStateList(pickIntegerInRange(0, 25));
+  return {id, url, description, likes, comments};
+};
+
+
+/**
+ * создаст список объектов фотографий
+ * @param {number} length
+ * @return {ImageState[]}
+ */
+const createImageStateList = (length = 25) => {
+  const list = new Array(length).fill(1);
+
+  return list.map((start, index) => createImageState(start + index));
+};
+
+export default createImageStateList;
